+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### New Features
+
+* Add parsing of `PivotCache` list
+
 ## 0.6.0 (2020-05-29)
 
 ### New Features